requirements.txt
pytest-html

pytest-xdist

pytest-ordering

pytest-rerunfailures

allure-pytest